Fix the bug with truncation error in LInearPropagators.jl
Due to the truncation error, which appears because of the Float32 precision, the diffraction in free space was simulated with errors. The truncation error appeared in calculation of exponent, exp(-1im * KZ * z), because of the large phase values. As the solution, we reduce the value of KZ by going to a moving reference frame: KZ = KZ - w0 / vf.