Faster Hankel transform for RT geometry.
The spectrum of an analytic function is nonzero only for positive frequencies. Due to the way how FFT store arrays, the positive frequency part of the spectrum occupies only the first half of the array. Therefore, Hankel transform of the spatio-temporal spectrum of an analytic function can be speed up by applying it only to the first half of the input 2D array.